# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Drogaria",
        "description": "Esta documentação descreve os endpoints disponíveis para integração com o sistema web da Drogaria, desenvolvido em Laravel. A API pode ser utilizada por qualquer linguagem de programação capaz de realizar requisições HTTP.",
        "contact": {
            "name": "Bruno",
            "email": "bruno.facomp2016@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/vendedores": {
            "get": {
                "tags": [
                    "Vendedores"
                ],
                "summary": "Retorna a lista de todos os vendedores cadastrados",
                "operationId": "c102ef2bdf28c3039d3de106f77ae7f5",
                "responses": {
                    "200": {
                        "description": "Lista completa de vendedores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nome": {
                                                "type": "string",
                                                "example": "Maria dos Santos"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-01-01T12:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-01-02T12:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Vendedores"
                ],
                "summary": "Cria um novo vendedor",
                "operationId": "bd6755f9235b4179db7fadf66b99743b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Maria dos Santos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vendedor criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Maria dos Santos"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-01T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-01T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/vendedores/{id}": {
            "get": {
                "tags": [
                    "Vendedores"
                ],
                "summary": "Consulta um vendedor específico pelo ID",
                "operationId": "d3da8de5be282fca5c53d06699979b32",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador único do vendedor",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do vendedor encontrados com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Maria dos Santos"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-01T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-02T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vendedor não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vendedores"
                ],
                "summary": "Atualiza os dados de um vendedor existente",
                "operationId": "6f7463eaca58b7eadc178613afe77e77",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identificador do vendedor",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Maria dos Santos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vendedor atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Maria dos Santos"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-01T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-01-02T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vendedor não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vendedores"
                ],
                "summary": "Exclui um vendedor pelo ID",
                "operationId": "76c93dacd251ca61388366d1b83d8300",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do vendedor a ser removido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Vendedor excluído com sucesso"
                    },
                    "404": {
                        "description": "Vendedor não encontrado"
                    }
                }
            }
        }
    },
...
}